requirements.txt
anyio==3.6.1
async-timeout==4.0.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.0
packaging==21.3
pycodestyle==2.9.1
pycryptodome==3.15.0
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
